+++ include/xieite/geometry/getIntersection.hpp
@@ -10,7 +10,7 @@ namespace xieite::geometry {
 	template<std::derived_from<xieite::geometry::LineLike> L, std::derived_from<xieite::geometry::LineLike> M>
 	[[nodiscard]]
 	constexpr std::optional<xieite::geometry::Point> getIntersection(const L& lineLike1, const M& lineLike2) noexcept {
-		const long double a = (lineLike1.start.x - lineLike1.end.x) * (lineLike2.start.y - lineLike2.end.y) - (lineLike1.start.y - lineLike1.end.y) * (lineLike2.start.x - lineLike2.end.x);
+		const double a = (lineLike1.start.x - lineLike1.end.x) * (lineLike2.start.y - lineLike2.end.y) - (lineLike1.start.y - lineLike1.end.y) * (lineLike2.start.x - lineLike2.end.x);
 		if (!xieite::math::approximatelyEqual(a, 0)) {
 			const xieite::geometry::Point intersection(((lineLike2.start.x - lineLike2.end.x) * (lineLike1.start.x * lineLike1.end.y - lineLike1.start.y * lineLike1.end.x) - (lineLike1.start.x - lineLike1.end.x) * (lineLike2.start.x * lineLike2.end.y - lineLike2.start.y * lineLike2.end.x)) / a, ((lineLike2.start.y - lineLike2.end.y) * (lineLike1.start.x * lineLike1.end.y - lineLike1.start.y * lineLike1.end.x) - (lineLike1.start.y - lineLike1.end.y) * (lineLike2.start.x * lineLike2.end.y - lineLike2.start.y * lineLike2.end.x)) / a);
 			if (xieite::geometry::containsPoint(lineLike1, intersection) && xieite::geometry::containsPoint(lineLike2, intersection))
